[ALS-5514] Change POST to GET in OktaAuthenticationService
This update modifies the HTTP method for the authenticate method in OktaAuthenticationService from POST to GET. This change aligns with our data retrieval patterns, and adapts it to receive parameters from the URL instead of the request body. The import statement also refactored to use wildcard import for javax.ws.rs package.