[ALS-5979] AIM-AHEAD PIC-SURE: Downloaded dataset seems incorrect (#161)
* Refactor FENCEAuthenticationService to customize query fields

Different fields for queryTemplateText are set based on the idp_provider in FENCEAuthenticationService. The "?fields" parameter of the query is now conditionally filled, with it being set to the parentAccessionField if idp_provider equals "fence", and empty otherwise. This change provides more flexibility for FENCE authentication depending on the provider setting.

* Remove empty string for fields

* Flip equals to avoid npe