Refactor TokenService to use CustomApplicationDetails
In the TokenService, replaced the direct retrieval of the Application principal with the use of the newly introduced CustomApplicationDetails object. This approach provides a more structured way to handle application data. Also, refactored logging statements to use placeholders, thus improving the readability of the code and efficiency of string concatenation.