[ALS-6102] Update AccessRule entity relationships and remove unneeded properties
The 'AccessRule' entity has been updated to use a Many-To-Many relationship for 'subAccessRule', with a migrated association table 'accessRule_subRule'. Furthermore, unnecessary JsonIgnore annotations have been removed for 'mergedValues' and 'mergedName' properties. The 'subAccessRuleParent' property and its related getters and setters are also removed as they are no longer required.