[ALS-6102] Remove circular dependency in access_rule table
A new SQL migration script has been added to remove the circular dependency in the access_rule table and replace it with a reference to the join table, accessRule_subRule. The script inserts data from the circular dependency into the join table, updates references in the original table to point to the join table, and finally drops the dependency column from the original table.