Refactor FENCEAuthenticationService and update logging
The over arching goal of this commit is to increase performance of the fence login code. In this commit I have reduced the number of times the database is queried. Instead of checking the database for each individual role we are checking for ALL roles that exists in the database using a single "WHERE role.name IN" query. I use the returned list to determine which roles don't exist in the database. We then only create those roles. Those roles are now batch inserted.