Refactor FENCEAuthenticationService and RoleRepository
Implemented an extensive refactor of FENCEAuthenticationService's user role management. Updated functionalities include: removal of roles not in the project_access_set, detection of previously assigned roles, assigning database-existing roles to users, and creation and upserting of non-existing roles in the database. Corrected the logic to re-assign new roles to the user after persistence of these roles. Adjustments also made in RoleRepository to return roles by names.