Refactor FENCE authentication to utilize StudyMetaData model
Introduced the use of StudyMetaData model in FENCE authentication for increased type safety and enforcement of encapsulation, improving the robustness of the code. The changes provide a more structured approach to data management, replacing previous Map implementations with the StudyMetaData model. The JSON files are now deserialized into the BioDataCatalyst model which contains a list of StudyMetaData.