Update configuration and Dockerfile
The psama.env configuration file and the bdc.Dockerfile have been updated. In psama.env, the CLIENT_SECRET has been nullified for security reasons. In the Dockerfile, the file name for the additional configuration and the value for SPRING_CONFIG_ADDITIONAL_LOCATION environment variable have been changed to suit the new database configurations.