Refactor FENCE service initialization and role creation
This update simplifies the initialization of the FENCE service in the FENCEAuthenticationService class by using the @PostConstruct annotation. It also moves the creation and persistence of access roles based on the FENCE mapping to the FenceMappingUtility class using the @EventListener(ContextRefreshedEvent.class) annotation. This improves code organization and ensures the correct sequence of operations.